+++ _is_override_code.d.ts
@@ -1,0 +1,2 @@
+export declare function _is_override_code(infostring: string): boolean;
+export { _is_override_code as _is__code__override };