+++ _markdown_html.d.ts
@@ -1,0 +1,5 @@
+/**
+ * Returns html from the given markdown
+ */
+export declare function _markdown_html(markdown: any): any;
+export { _markdown_html as _html__markdown };