+++ _markdown_html.ts
@@ -1,0 +1,20 @@
+import marked from 'marked'
+import { _is_override_code } from './_is_override_code'
+/**
+ * Returns html from the given markdown
+ */
+export function _markdown_html(markdown) {
+	const renderer = new marked.Renderer()
+	renderer.code = code__override
+	return marked(markdown, { renderer })
+	function code__override(code, infostring) {
+		return (
+			_is_override_code(infostring)
+			? ''
+			: code
+		)
+	}
+}
+export {
+	_markdown_html as _html__markdown
+}