+++ _markdown_name_a1.d.ts
@@ -1,0 +1,6 @@
+import type { markdown_opts_type } from './markdown_opts_type';
+/**
+ * Returns an array of names for each markdown file in opts
+ */
+export declare function _markdown_name_a1(opts: markdown_opts_type): Promise<string[] | undefined>;
+export { _markdown_name_a1 as _a1__name };