+++ _resolve_md_file_path_txt.d.ts
@@ -1,0 +1,2 @@
+export declare function _resolve_md_file_path_txt(path_txt: string): Promise<string | false | undefined>;
+export { _resolve_md_file_path_txt as _txt__path__file__md__resolve };