+++ markdown_opts_type.ts
@@ -1,0 +1,7 @@
+/**
+ * Name for a markdown file
+ * @typedef {string} name__md
+ */
+export interface markdown_opts_type {
+	dir:string
+}